+++ src/main/java/com/techfork/auth/infrastructure/kakao/response/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.techfork.auth.infrastructure.kakao.dto;
+package com.techfork.auth.infrastructure.kakao.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
